# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MegaOctav API",
        "description": "Dokumentasi API untuk sistem reservasi MegaOctav",
        "contact": {
            "email": "imeg290@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/films": {
            "get": {
                "tags": [
                    "Film"
                ],
                "summary": "List all films",
                "operationId": "11a3cd6ea05ca00cbbf1e4eadbcfb0ad",
                "responses": {
                    "200": {
                        "description": "List of films",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Film"
                ],
                "summary": "Add a new film",
                "operationId": "e39004ef3e890f2656092b9e5e1c1b75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "film_title",
                                    "synopsis",
                                    "genre",
                                    "duration",
                                    "rating_film"
                                ],
                                "properties": {
                                    "film_title": {
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "type": "string"
                                    },
                                    "genre": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer"
                                    },
                                    "rating_film": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film created"
                    }
                }
            }
        },
        "/films/{id}": {
            "get": {
                "tags": [
                    "Film"
                ],
                "summary": "Get film by ID",
                "operationId": "b65fefef4f94418153cfac8a3cc2563d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film data"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Film"
                ],
                "summary": "Update film data",
                "operationId": "a3c230efffe40d6b03e365d02972f7a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "film_title": {
                                        "type": "string"
                                    },
                                    "synopsis": {
                                        "type": "string"
                                    },
                                    "genre": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer"
                                    },
                                    "rating_film": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Film"
                ],
                "summary": "Delete a film",
                "operationId": "c3c42175bee8509a216ae73852b1fe2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film deleted"
                    }
                }
            }
        },
...
        "/seats": {
            "post": {
                "tags": [
                    "Seat"
                ],
                "summary": "Create new seat",
                "operationId": "c1da66ed128798a203d7cf9957ef7e36",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id",
                                    "number",
                                    "status_seats"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status_seats": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "booked"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Seat created"
                    }
                }
            }
        },
        "/seats/{id}": {
            "get": {
                "tags": [
                    "Seat"
                ],
                "summary": "Get seat by ID",
                "operationId": "9e46d09f564c08ab96bd419b0dc1ee3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seat found"
                    },
                    "404": {
                        "description": "Seat not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Seat"
                ],
                "summary": "Update a seat",
                "operationId": "389de4043141c5b6aac8de6848819c81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status_seats": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "booked"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Seat updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Seat"
                ],
                "summary": "Delete a seat",
                "operationId": "b78f06764fb1f63d6f1db12154b5d488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seat deleted"
                    }
                }
            }
        },
...
    },
...
}